# packages/api/openapi.yaml
openapi: 3.0.0
info:
  title: 1BRC for traP API
  version: 0.0.0
...
paths:
  /api/v1/access-key:
    post:
      operationId: issueAccessKey
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessKeyResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
    delete:
      operationId: revokeAccessKey
      parameters: []
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
  /api/v1/admin/datasets/import:
    post:
      operationId: importDatasets
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                    format: int32
                required:
                  - imported
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetManifest'
  /api/v1/admin/private/publish:
    post:
      operationId: publishPrivate
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  published:
                    type: boolean
                    enum:
                      - true
                required:
                  - published
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
  /api/v1/admin/private/unpublish:
    post:
      operationId: unpublishPrivate
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  published:
                    type: boolean
                    enum:
                      - false
                required:
                  - published
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
  /api/v1/admin/submissions:
    get:
      operationId: listAdminSubmissions
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionsResponse'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '403':
          description: Access is forbidden.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    $ref: '#/components/schemas/ErrorDetail'
                required:
                  - error
...
components:
  schemas:
    AccessKeyResponse:
      type: object
      required:
        - accessKey
        - prefix
      properties:
        accessKey:
          type: string
        prefix:
          type: string
...
    DatasetArtifact:
      type: object
      required:
        - id
        - kind
        - label
        - objectKey
        - rows
        - compressedBytes
        - uncompressedBytes
        - compressedSha256
        - uncompressedSha256
        - isPublic
      properties:
        id:
          type: string
          pattern: ^[a-z0-9][a-z0-9-]{0,63}$
        kind:
          $ref: '#/components/schemas/DatasetKind'
        label:
          type: string
          minLength: 1
          maxLength: 128
        objectKey:
          type: string
          maxLength: 1024
          pattern: ^[\u0020-\u007E]+$
        rows:
          type: integer
          format: int64
          minimum: 1
        compressedBytes:
          type: integer
          format: int64
          minimum: 1
        uncompressedBytes:
          type: integer
          format: int64
          minimum: 1
        compressedSha256:
          type: string
          pattern: ^[a-f0-9]{64}$
        uncompressedSha256:
          type: string
          pattern: ^[a-f0-9]{64}$
        isPublic:
          type: boolean
...
    DatasetManifest:
      type: object
      required:
        - schemaVersion
        - contestId
        - generatedAt
        - generatorRevision
        - artifacts
      properties:
        schemaVersion:
          type: number
          enum:
            - 1
        contestId:
          type: string
          pattern: ^[a-z0-9][a-z0-9-]{0,127}$
        generatedAt:
          $ref: '#/components/schemas/DateTimeString'
        generatorRevision:
          type: string
          minLength: 1
          maxLength: 128
        artifacts:
          type: array
          items:
            $ref: '#/components/schemas/DatasetArtifact'
          minItems: 4
...
    DateTimeString:
      type: string
    ErrorDetail:
      type: object
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
...
    SubmissionsResponse:
      type: object
      required:
        - submissions
      properties:
        submissions:
          type: array
          items:
            $ref: '#/components/schemas/Submission'
...